refactor(mcp): organize tool handlers one domain per module
Make the MCP handler layer uniform and codify the convention so it stays
uniform (mirroring data-access's "one repository per aggregate" guard):

- Rename tools.ts -> memory-tools.ts (buildHandlers -> buildMemoryHandlers,
  ToolDeps -> MemoryToolDeps).
- Split the 912-line sessions-tools.ts into session-tools.ts,
  prompt-tools.ts, and observability-tools.ts; context/timeline move into
  memory-tools.ts.
- Centralize shared helpers in _shared.ts (scopeFromContext, routerKey,
  clamp, snippet, serializeMemory, resolveSessionId) and move errToMcp into
  errors.ts; drop the duplicate routerKey from project-tools.ts.
- Add a grep invariant (invariants.test.ts): no tools.ts; every *-tools.ts
  exports exactly one build*Handlers; errToMcp/routerKey defined once.

No tool is added, removed, renamed, or has its contract changed; server.ts
registers the same tool surface. Spec deltas under
openspec/changes/reorganize-mcp-tools (mcp-api ADDED requirement).
