+++ apps/server/src/mcp/index.ts
@@ -6,4 +6,4 @@ export { createMcpServer } from './server.js';
 export type { CreateMcpServerOptions } from './server.js';
 export { McpTransportManager } from './transport.js';
 export { mcpError } from './errors.js';
-export { buildHandlers } from './tools.js';
+export { buildMemoryHandlers } from './memory-tools.js';
